requirements.txt
ultralytics
opencv-python
numpy
lap
pillow
